techDebt(removeOldBulkUpload): Delete all unused bulk-upload-v1 files… (#3151)
* techDebt(removeOldBulkUpload): Delete all unused bulk-upload-v1 files and remove conditional bulk upload routing from app-routing.module

* techDebt(removeOldBulkUpload): Delete all unused bulk-upload-v1 files and remove conditional bulk upload routing from app-routing.module

* techDebt(removeOldBulkUpload): Change bulk-upload-v2 folder name to bulk-upload and remove v2s from component names

* techDebt(removeOldBulkUpload): Add import statement to fix error report tests

* techDebt(removeOldBulkUpload): Revert bulk upload page to import BulkUploadService and not V2

* techDebt(removeOldBulkUpload): Delete references-created-page component

* techDebt(removeOldBulkUpload): Provide BulkUploadServiceV2 for bulk-upload-page component to fix endpoints bug

* techDebt(removeOldBulkUpload): Change folder name back to bulk-upload-v2

* techDebt(removeOldBulkUpload): Fix import statements
