+++ src/app/app-routing.module.ts
@@ -1,9 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { PageNotFoundComponent } from '@core/components/error/page-not-found/page-not-found.component';
-import {
-  ProblemWithTheServiceComponent,
-} from '@core/components/error/problem-with-the-service/problem-with-the-service.component';
+import { ProblemWithTheServiceComponent } from '@core/components/error/problem-with-the-service/problem-with-the-service.component';
 import { AuthGuard } from '@core/guards/auth/auth.guard';
 import { LoggedOutGuard } from '@core/guards/logged-out/logged-out.guard';
 import { MigratedUserGuard } from '@core/guards/migrated-user/migrated-user.guard';
@@ -18,37 +16,9 @@ import { DashboardComponent } from '@features/dashboard/dashboard.component';
 import { ForgotYourPasswordComponent } from '@features/forgot-your-password/forgot-your-password.component';
 import { LoginComponent } from '@features/login/login.component';
 import { LogoutComponent } from '@features/logout/logout.component';
-import {
-  MigratedUserTermsConditionsComponent,
-} from '@features/migrated-user-terms-conditions/migrated-user-terms-conditions.component';
+import { MigratedUserTermsConditionsComponent } from '@features/migrated-user-terms-conditions/migrated-user-terms-conditions.component';
 import { ResetPasswordComponent } from '@features/reset-password/reset-password.component';
 import { SatisfactionSurveyComponent } from '@features/satisfaction-survey/satisfaction-survey.component';
-
-import { environment } from '../environments/environment';
-
-const bulkUploadRouting = {
-  other: {
-    path: 'bulk-upload',
-    loadChildren: () => import('@features/bulk-upload-v2/bulk-upload.module').then((m) => m.BulkUploadV2Module),
-    data: {
-      title: 'Bulk Upload',
-    },
-  },
-  test: {
-    path: 'bulk-upload',
-    loadChildren: () => import('@features/bulk-upload-v2/bulk-upload.module').then((m) => m.BulkUploadV2Module),
-    data: {
-      title: 'Bulk Upload',
-    },
-  },
-  production: {
-    path: 'bulk-upload',
-    loadChildren: () => import('@features/bulk-upload-v2/bulk-upload.module').then((m) => m.BulkUploadV2Module),
-    data: {
-      title: 'Bulk Upload',
-    },
-  },
-}[environment.environmentName];
 
 const routes: Routes = [
   {
@@ -156,7 +126,11 @@ const routes: Routes = [
         component: DashboardComponent,
         data: { title: 'Dashboard' },
       },
-      bulkUploadRouting,
+      {
+        path: 'bulk-upload',
+        loadChildren: () => import('@features/bulk-upload-v2/bulk-upload.module').then((m) => m.BulkUploadModule),
+        data: { title: 'Bulk Upload' },
+      },
       {
         path: 'search-users',
         loadChildren: () => import('@features/search/search.module').then((m) => m.SearchModule),
