+++ src/app/core/guards/bulk-upload/bulk-upload-missing.guard.ts
@@ -22,15 +22,20 @@ export class BulkUploadMissingGuard implements CanActivate {
       ? this.establishmentService.primaryWorkplace.uid
       : this.establishmentService.establishmentId;
 
-
-
     return this.bulkUploadService.getMissingRef(primaryWorkplaceID).pipe(
       map((response) => {
-        if ( response.establishmentList.some(establishment => this.adminSkipService.skippedWorkplaces.includes(establishment.uid) )) {
+        if (
+          response.establishmentList.some((establishment) =>
+            this.adminSkipService.skippedWorkplaces.includes(establishment.uid),
+          )
+        ) {
           return true;
         }
 
-        if ((response.establishment > 0 && this.adminSkipService.skipWorkplaceReferences !== true) || response.worker > 0) {
+        if (
+          (response.establishment > 0 && this.adminSkipService.skipWorkplaceReferences !== true) ||
+          response.worker > 0
+        ) {
           const redirect: UrlTree = this.router.parseUrl('/bulk-upload/missing');
           return redirect;
         }
