+++ src/app/features/bulk-upload-v2/bulk-upload-references/missing-staff-references/missing-staff-references-page.component.spec.ts
@@ -20,7 +20,7 @@ import { MockBulkUploadService } from '@core/test-utils/MockBulkUploadService';
 import { MockEstablishmentService } from '@core/test-utils/MockEstablishmentService';
 import { MockWorkerService, workerBuilder } from '@core/test-utils/MockWorkerService';
 import { AdminSkipService } from '@features/bulk-upload-v2/admin-skip.service';
-import { BulkUploadV2Module } from '@features/bulk-upload-v2/bulk-upload.module';
+import { BulkUploadModule } from '@features/bulk-upload-v2/bulk-upload.module';
 import { SharedModule } from '@shared/shared.module';
 import { fireEvent, render } from '@testing-library/angular';
 import { Observable, Subject } from 'rxjs';
@@ -30,7 +30,7 @@ import { MissingStaffReferencesComponent } from './missing-staff-references-page
 describe('MissingStaffReferencesComponent', () => {
   async function setup(references: Worker[] = []) {
     const component = await render(MissingStaffReferencesComponent, {
-      imports: [SharedModule, RouterModule, RouterTestingModule, HttpClientTestingModule, BulkUploadV2Module],
+      imports: [SharedModule, RouterModule, RouterTestingModule, HttpClientTestingModule, BulkUploadModule],
       providers: [
         {
           provide: EstablishmentService,
