+++ src/app/features/bulk-upload-v2/bulk-upload-routing.module.ts
@@ -10,14 +10,13 @@ import { MissingWorkplacesReferencesResolver } from '@core/resolvers/missing-wor
 import { StaffReferencesResolver } from '@core/resolvers/staff-references.resolver';
 import { WorkplacesReferencesResolver } from '@core/resolvers/workplace-references.resolver';
 import { AboutBulkUploadComponent } from '@features/bulk-upload-v2/about-bulk-upload/about-bulk-upload.component';
-import { BulkUploadPageV2Component } from '@features/bulk-upload-v2/bulk-upload-page/bulk-upload-page.component';
 import { MissingStaffReferencesComponent } from '@features/bulk-upload-v2/bulk-upload-references/missing-staff-references/missing-staff-references-page.component';
 import { MissingWorkplaceReferencesComponent } from '@features/bulk-upload-v2/bulk-upload-references/missing-workplace-references/missing-workplace-references-page.component';
-import { LastBulkUploadComponent } from '@features/bulk-upload-v2/last-bulk-upload/last-bulk-upload.component';
 import { BulkUploadStartPageComponent } from '@features/bulk-upload-v2/bulk-upload-start-page/bulk-upload-start-page.component';
-import { ReferencesCreatedPageComponent } from '@features/bulk-upload/references-created-page/references-created-page.component';
+import { LastBulkUploadComponent } from '@features/bulk-upload-v2/last-bulk-upload/last-bulk-upload.component';
 
 import { BulkUploadMissingPageComponent } from './bulk-upload-missing/bulk-upload-missing-page.component';
+import { BulkUploadPageComponent } from './bulk-upload-page/bulk-upload-page.component';
 import { StaffReferencesComponent } from './bulk-upload-references/staff-references/staff-references-page.component';
 import { WorkplaceReferencesComponent } from './bulk-upload-references/workplace-references/workplace-references-page.component';
 import { ErrorPageComponent } from './error-page/error-page.component';
@@ -25,7 +24,7 @@ import { ErrorPageComponent } from './error-page/error-page.component';
 const routes: Routes = [
   {
     path: '',
-    component: BulkUploadPageV2Component,
+    component: BulkUploadPageComponent,
     canActivate: [BulkUploadStartGuard, BulkUploadMissingGuard],
     data: { title: 'Home' },
   },
@@ -85,11 +84,6 @@ const routes: Routes = [
     data: { title: 'Staff references' },
   },
   {
-    path: 'workplace-and-staff-references/success',
-    component: ReferencesCreatedPageComponent,
-    data: { title: 'Workplace and staff references complete' },
-  },
-  {
     path: 'error-report',
     component: ErrorPageComponent,
     data: { title: 'Error Report' },
@@ -101,4 +95,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
-export class BulkUploadV2RoutingModule {}
+export class BulkUploadRoutingModule {}
