+++ src/app/features/bulk-upload-v2/bulk-upload-sidebar/bulk-upload-related-content/bulk-upload-related-content.component.spec.ts
@@ -3,7 +3,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AuthService } from '@core/services/auth.service';
 import { MockAuthService } from '@core/test-utils/MockAuthService';
-import { BulkUploadV2Module } from '@features/bulk-upload-v2/bulk-upload.module';
+import { BulkUploadModule } from '@features/bulk-upload-v2/bulk-upload.module';
 import { render } from '@testing-library/angular';
 
 import { BulkUploadRelatedContentComponent } from './bulk-upload-related-content.component';
@@ -11,7 +11,7 @@ import { BulkUploadRelatedContentComponent } from './bulk-upload-related-content
 describe('BulkUploadRelatedContentComponent', () => {
   const setup = async (isAdmin = false, isLoggedIn: boolean = true) => {
     const { fixture, getByText, queryByText } = await render(BulkUploadRelatedContentComponent, {
-      imports: [RouterTestingModule, HttpClientTestingModule, BrowserModule, BulkUploadV2Module],
+      imports: [RouterTestingModule, HttpClientTestingModule, BrowserModule, BulkUploadModule],
       providers: [
         {
           provide: AuthService,
