+++ src/app/features/bulk-upload-v2/error-page/error-page.component.spec.ts
@@ -1,27 +1,27 @@
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 import { BrowserModule } from '@angular/platform-browser';
+import { ActivatedRoute } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
-import { BulkUploadService } from '@core/services/bulk-upload.service';
+import { BreadcrumbService } from '@core/services/breadcrumb.service';
+import { BulkUploadService, BulkUploadServiceV2 } from '@core/services/bulk-upload.service';
 import { EstablishmentService } from '@core/services/establishment.service';
-import { MockBulkUploadService, errorReport } from '@core/test-utils/MockBulkUploadService';
+import { MockBreadcrumbService } from '@core/test-utils/MockBreadcrumbService';
 import { MockEstablishmentService } from '@core/test-utils/MockEstablishmentService';
+import { MockBulkUploadService, errorReport } from '@core/test-utils/MockBulkUploadService';
 import { render } from '@testing-library/angular';
 
-import { BulkUploadV2Module } from '../bulk-upload.module';
+import { BulkUploadModule } from '../bulk-upload.module';
 import { ErrorPageComponent } from './error-page.component';
-import { ActivatedRoute } from '@angular/router';
-import { BreadcrumbService } from '@core/services/breadcrumb.service';
-import { MockBreadcrumbService } from '@core/test-utils/MockBreadcrumbService';
 
 describe('ErrorPageComponent', () => {
   const getErrorPageComponent = async () => {
     return await render(ErrorPageComponent, {
-      imports: [RouterTestingModule, HttpClientTestingModule, BrowserModule, BulkUploadV2Module],
+      imports: [RouterTestingModule, HttpClientTestingModule, BrowserModule, BulkUploadModule],
       declarations: [ErrorPageComponent],
       providers: [
         { provide: EstablishmentService, useClass: MockEstablishmentService },
-        { provide: BulkUploadService, useClass: MockBulkUploadService },
+        { provide: BulkUploadService, useClass: BulkUploadServiceV2 },
         { provide: BreadcrumbService, useClass: MockBreadcrumbService },
 
         {
