--- src/app/features/bulk-upload/bulk-upload-page/bulk-upload-page.component.html
@@ -1,33 +1,0 @@
-<app-error-summary
-  *ngIf="showErrorSummary"
-  [customErrors]="uploadValidationErrors"
-  [serverError]="serverError"
-  [formErrorsMap]="formErrorsMap"
-  [form]="form"
-></app-error-summary>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-l">
-      <span *ngIf="establishment?.name" class="govuk-caption-xl">{{ establishment?.name }}</span>
-      Bulk upload
-    </h1>
-    <p>Bulk upload is used to add and edit large quantities of staff, workplace and training information.</p>
-  </div>
-</div>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-full">
-    <ol class="govuk-list govuk-list--number govuk-!-margin-top-8 govuk-!-padding-left-5">
-      <li class="govuk-!-padding-left-5 govuk-!-font-size-24 govuk-!-padding-bottom-6 govuk-util__bold">
-        <app-check-workplace-references class="govuk-util__block"></app-check-workplace-references>
-      </li>
-      <li class="govuk-!-padding-left-5 govuk-!-font-size-24 govuk-!-padding-bottom-6 govuk-util__bold">
-        <app-download-data-files class="govuk-util__block"></app-download-data-files>
-      </li>
-      <li class="govuk-!-padding-left-5 govuk-!-font-size-24 govuk-!-padding-bottom-6 govuk-util__bold">
-        <app-upload-data-files class="govuk-util__block"></app-upload-data-files>
-      </li>
-    </ol>
-  </div>
-</div>
