--- src/app/features/bulk-upload/bulk-upload-references/bulk-upload-references.html
@@ -1,147 +1,0 @@
-<app-error-summary
-  *ngIf="(submitted && form.invalid) || serverError"
-  [serverError]="serverError"
-  [formErrorsMap]="formErrorsMap"
-  [form]="form"
-></app-error-summary>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-l">
-      <span *ngIf="primaryEstablishmentName" class="govuk-caption-xl">{{ primaryEstablishmentName }}</span>
-      {{ referenceType }} references
-    </h1>
-    <p>{{ referenceTypeInfo }}</p>
-  </div>
-</div>
-
-<div class="govuk-grid-row" *ngIf="renderForm">
-  <ng-container *ngIf="referenceType === referenceTypeEnum.Worker && establishmentName && primaryWorkplace.isParent">
-    <div class="govuk-grid-column-two-thirds-from-desktop">
-      <h2 class="govuk-heading-m asc__reference-subsidiary-name govuk-!-margin-0">{{ establishmentName }}</h2>
-    </div>
-    <div *ngIf="remainingEstablishments > 0" class="govuk-grid-column-one-third-from-desktop">
-      <p class="govuk-!-margin-0 govuk-util__align-right">
-        <strong>{{ remainingEstablishments }}</strong>
-        {{ remainingEstablishments | i18nPlural: { '=1': 'workplace', other: 'workplaces' } }} remaining
-      </p>
-    </div>
-  </ng-container>
-  <div class="govuk-grid-column-full">
-    <form
-      #formEl
-      *ngIf="references.length"
-      class="govuk-!-margin-top-6"
-      novalidate
-      (ngSubmit)="onSubmit(true)"
-      [formGroup]="form"
-      id="server-error"
-    >
-      <div class="govuk__grid govuk__grid-container">
-        <p class="govuk-!-margin-bottom-0 govuk__grid-column-1">
-          <strong>{{ columnOneLabel }}</strong>
-        </p>
-        <p class="govuk-!-margin-bottom-0 govuk__grid-column-2">
-          <strong>{{ columnTwoLabel }}</strong>
-        </p>
-      </div>
-
-      <div
-        class="govuk-form-group govuk__grid govuk__grid-container"
-        *ngFor="let reference of references"
-        [class.govuk-form-group--error]="form.get('reference-' + reference.uid).errors && submitted"
-      >
-        <label class="govuk-label govuk-!-margin-bottom-0 govuk__grid-column-1" for="reference-{{ reference.uid }}">
-          {{ reference.name ? reference.name : reference.nameOrId }}
-        </label>
-        <span
-          id="reference-{{ reference.uid }}-error"
-          class="govuk-error-message govuk__grid-row-start-1 asc__reference-grid-error"
-          *ngIf="form.get('reference-' + reference.uid).errors && submitted"
-        >
-          <span class="govuk-visually-hidden">Error:</span> {{ getFirstErrorMessage('reference-' + reference.uid) }}
-        </span>
-        <input
-          class="govuk-input govuk__grid-column-2"
-          [class.govuk-input--error]="form.get('reference-' + reference.uid).errors && submitted"
-          id="reference-{{ reference.uid }}"
-          name="reference-{{ reference.uid }}"
-          type="text"
-          [formControlName]="'reference-' + reference.uid"
-        />
-        <p
-          class="govuk-!-margin-bottom-0 govuk__justify-self-end govuk__grid-column-3"
-          *ngIf="referenceType === referenceTypeEnum.Establishment && reference.localIdentifier"
-        >
-          <a [routerLink]="['/bulk-upload/staff-references', reference.uid]" (click)="setReturn()">
-            <span class="govuk-visually-hidden">{{ reference.name }}</span>
-            Staff references
-          </a>
-        </p>
-      </div>
-    </form>
-
-    <div class="govuk-grid-row" *ngIf="!references.length">
-      <div class="govuk-grid-column-two-thirds-from-desktop">
-        <ng-container *ngIf="referenceType === referenceTypeEnum.Establishment; else staffRecordsNextSteps">
-          <p class="govuk-!-margin-top-4">There are no workplaces editable</p>
-        </ng-container>
-        <ng-template #staffRecordsNextSteps>
-          <p class="govuk-!-margin-top-4">
-            {{ referenceType }} records need to be added for this workplace before you can create {{
-            referenceType.toLowerCase() }} references.
-          </p>
-          <h3 class="govuk-heading-m">What you can do next?</h3>
-          <ul class="govuk-list govuk-list--bullet">
-            <li>
-              <a [routerLink]="addStaffRoute.url" [fragment]="addStaffRoute.fragment">
-                Add staff records <span class="govuk-visually-hidden">for {{ establishmentName }}</span>
-              </a>
-            </li>
-            <li>Continue without adding staff records</li>
-          </ul>
-        </ng-template>
-        <div class="govuk-button-group">
-          <a
-            role="button"
-            draggable="false"
-            class="govuk-button govuk-!-margin-right-8 govuk-!-margin-bottom-0"
-            [routerLink]="nextRoute"
-          >
-            {{ !!returnTo ? 'Return' : 'Continue' }}
-          </a>
-          <span class="govuk-visually-hidden">or</span>
-          <a
-            role="button"
-            draggable="false"
-            class="govuk-button govuk-button--link govuk-!-margin-bottom-0"
-            [routerLink]="['/bulk-upload/workplace-references']"
-          >
-            Exit
-          </a>
-        </div>
-      </div>
-    </div>
-
-    <div class="govuk-grid-row" *ngIf="references.length">
-      <div class="govuk-grid-column-full">
-        <div class="govuk-button-group">
-          <button
-            type="button"
-            class="govuk-button govuk-!-margin-right-8 govuk-!-margin-bottom-0"
-            (click)="onSubmit($event, true)"
-          >
-            {{ !!returnTo ? 'Save and return' : 'Save and continue' }}
-          </button>
-          <span class="govuk-visually-hidden">or</span>
-          <a
-            href="#"
-            class="govuk-list govuk-list--inline govuk-util__float-right govuk-!-margin-top-2"
-            (click)="onSubmit($event,false)"
-            >Save and exit</a
-          >
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
