--- src/app/features/bulk-upload/bulk-upload-references/bulk-upload-references.scss
@@ -1,9 +1,0 @@
-.asc__reference-grid-error {
-  grid-column-start: 1;
-  grid-column-end: 4;
-}
-
-.govuk__reference-subsidiary-name {
-  grid-column-start: 1;
-  grid-column-end: 3;
-}
