--- src/app/features/bulk-upload/bulk-upload-start-page/bulk-upload-start-page.component.html
@@ -1,36 +1,0 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds-from-desktop">
-    <h1 class="govuk-heading-xl">Bulk Upload</h1>
-    <p>You're about to begin bulk uploading files for the first time.</p>
-    <p>
-      Before you can begin to bulk upload you must create unique references for each workplace and staff member. These
-      were previously known as Local identifiers.
-    </p>
-  </div>
-</div>
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds-from-desktop">
-    <div class="govuk-!-margin-top-6">
-      <a
-        role="button"
-        draggable="false"
-        class="govuk-button govuk-button--start"
-        [routerLink]="['/bulk-upload/workplace-references']"
-      >
-        Start now
-
-        <svg
-          class="govuk-button__start-icon"
-          xmlns="http://www.w3.org/2000/svg"
-          width="17.5"
-          height="19"
-          viewBox="0 0 33 40"
-          aria-hidden="true"
-          focusable="false"
-        >
-          <path fill="currentColor" d="M0 0h13l20 20-20 20H0l20-20z" />
-        </svg>
-      </a>
-    </div>
-  </div>
-</div>
