--- src/app/features/bulk-upload/bulk-upload-start-page/bulk-upload-start-page.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-bulk-upload-start-page',
-  templateUrl: './bulk-upload-start-page.component.html',
-})
-export class BulkUploadStartPageComponent {
-  constructor() {}
-}
