--- src/app/features/bulk-upload/check-workplace-references/check-workplace-references.component.html
@@ -1,7 +1,0 @@
-<h2 class="govuk-heading-m">Check your workplace references</h2>
-<p>Check your workplace and staff references to make sure they are the <br />same as the files you're uploading.</p>
-<p><a [routerLink]="['/bulk-upload/workplace-references']">View workplace references</a></p>
-<!--
-  TODO uncomment once relevant sections have been developed
-  <p *ngIf="***">Last updated on {{ lastUpdated | date: 'd MMMM y' }}</p>
--->
