--- src/app/features/bulk-upload/check-workplace-references/check-workplace-references.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-check-workplace-references',
-  templateUrl: './check-workplace-references.component.html'
-})
-export class CheckWorkplaceReferencesComponent {}
