--- src/app/features/bulk-upload/file-validate-status/file-validate-status.component.scss
@@ -1,16 +1,0 @@
-.govuk-file-validate-status--validating {
-  background: govuk-colour("orange");
-}
-
-.govuk-file-validate-status--pass,
-.govuk-file-validate-status--fail {
-  color: govuk-colour("white");
-}
-
-.govuk-file-validate-status--pass {
-  background: govuk-colour("green");
-}
-
-.govuk-file-validate-status--fail {
-  background: govuk-colour("red");
-}
