--- src/app/features/bulk-upload/files-upload-progress/files-upload-progress.component.html
@@ -1,3 +1,0 @@
-<p>
-  files-upload-progress works!
-</p>
