--- src/app/features/bulk-upload/files-upload-progress/files-upload-progress.component.ts
@@ -1,11 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-files-upload-progress',
-  templateUrl: './files-upload-progress.component.html',
-})
-export class FilesUploadProgressComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit() {}
-}
