--- src/app/features/bulk-upload/references-created-page/references-created-page.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-references-created-page',
-  templateUrl: './references-created-page.component.html',
-})
-export class ReferencesCreatedPageComponent {
-  constructor() {}
-}
