--- src/app/features/bulk-upload/selected-files-list/selected-files-list.component.html
@@ -1,16 +1,0 @@
-<table class="govuk-table govuk-!-margin-top-6" *ngIf="selectedFiles?.length">
-  <thead class="govuk-table__head">
-    <tr class="govuk-table__row">
-      <th class="govuk-table__header" scope="col">File name</th>
-      <th class="govuk-table__header" scope="col">File type</th>
-      <th class="govuk-table__header" scope="col">File size</th>
-    </tr>
-  </thead>
-  <tbody class="govuk-table__body">
-    <tr class="govuk-table__row" *ngFor="let file of selectedFiles">
-      <td class="govuk-table__cell govuk-!-font-weight-regular">{{ file.name }}</td>
-      <td class="govuk-table__cell">{{ getFileType(file.name) }}</td>
-      <td class="govuk-table__cell">{{ transformFileSize(file.size) }}</td>
-    </tr>
-  </tbody>
-</table>
