--- src/app/features/bulk-upload/upload-data-files/upload-data-files.component.html
@@ -1,9 +1,0 @@
-<h2 class="govuk-heading-m">Upload data files</h2>
-<p class="govuk-caption-m">20Mb maximum file size. Zip larger files</p>
-<ul class="govuk-list govuk-list--bullet">
-  <li>choose files to upload</li>
-  <li>check uploaded files</li>
-  <li>complete upload</li>
-</ul>
-
-<app-files-upload></app-files-upload>
