--- src/app/features/bulk-upload/upload-data-files/upload-data-files.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-upload-data-files',
-  templateUrl: './upload-data-files.component.html',
-})
-export class UploadDataFilesComponent {}
